On html: added a template for the rows to be added and deleted the dummy rows.
On styles:  added styles for row isRead toggle and delete buttons. Edited the fields placeholder color an font.

On script: created the addRow() function that adds a row using the template. The cycle that edits the textContent property works excepting for the read/unread button, a conditional to be added for that field that converts the boolean in the isRead property to the read / unread chain.