On html: changed the deleteBtn from Id to class, considering this type of element will be added with each line in table.
On Styles: updated the #deleteBtn selector to .deleteBtn.

On script: updated the CreateDOMFields function name to something more readable: createInputFields.

Added a conditional that selects the isRead key in myLibrary and shows read/unread on the book status .textContent, rather than true/false.

Added functionality to the delete button inside the addRow function.