On html: changed the attribute in the statusToggle button to a class, rather than an id.
On script: Updated the row class from a number to 'index#'

Added an index number to the statusToggle button being added.

Updated the addRow function to edit the statusToggle button content to Read/Unread, rather than updating the container of the button. This was deleting the button node.

Added an event listener on the statusBtn that toggles the Read/Unread status by changing the text content in the button node.